Add files via upload
About Turning Moment Calculation: The crank-pin circle radius of a horizontal engine is 0.3 m. The mass of the reciprocating parts is 250 kg. The net load acting on the piston is 68730 N. The connecting rod length between centers is 0.6 m and the engine runs at 250 r.p.m. Plot the turning moment diagram and find out the maximum torque.